+++ lib/screens/settings/getsandpostsofadmin/lookingforgetscreen.dart
@@ -77,6 +77,10 @@ class _LookingForGetScreenState extends ConsumerState<LookingForGetScreen> {
                               ),
                               child: ListTile(
                                 title: Text(item.value ?? ''),
+                                subtitle: Text(
+                                    item.mode?.value ?? '',
+                                    style: const TextStyle(color: Colors.grey),
+                                  ),
                                 trailing: Row(
                                   mainAxisSize: MainAxisSize.min,
                                   children: [
@@ -91,6 +95,7 @@ class _LookingForGetScreenState extends ConsumerState<LookingForGetScreen> {
                                           arguments: {
                                             'id': item.id,
                                             'value': item.value,
+                                            'modeid':item.mode?.id,
                                           },
                                         );
                                       },
@@ -98,16 +103,41 @@ class _LookingForGetScreenState extends ConsumerState<LookingForGetScreen> {
                                       
                                     ),
                                     IconButton(
-                                      icon: const Icon(Icons.delete,
-                                          color: Colors.red),
-                                      onPressed: ()async {
-                                        // TODO: Handle delete action
-                                        await ref.read(lookingProvider.notifier).deleteLookingfor(item.id);
-                                          ScaffoldMessenger.of(context).showSnackBar(
-                                            const SnackBar(content: Text("Deleted successfully")),
+                                    icon: const Icon(Icons.delete, color: Colors.red),
+                                    onPressed: () async {
+                                      final confirm = await showDialog<bool>(
+                                        context: context,
+                                        builder: (context) {
+                                          return AlertDialog(
+                                            title: const Text("Confirm Delete"),
+                                            content: const Text("Are you sure you want to delete this religion?"),
+                                            actions: [
+                                              TextButton(
+                                                onPressed: () => Navigator.pop(context, false), // Cancel
+                                                child: const Text("Cancel"),
+                                              ),
+                                              ElevatedButton(
+                                                style: ElevatedButton.styleFrom(
+                                                  backgroundColor: Colors.red,
+                                                ),
+                                                onPressed: () => Navigator.pop(context, true), // Confirm
+                                                child: const Text("Delete"),
+                                              ),
+                                            ],
                                           );
-                                      },
-                                    ),
+                                        },
+                                      );
+
+                                      if (confirm == true) {
+                                        await ref.read(lookingProvider.notifier).deleteLookingfor(item.id);
+                                        ScaffoldMessenger.of(context).showSnackBar(
+                                          const SnackBar(content: Text("Deleted successfully")),
+                                        );
+                                      }
+                                    },
+                                  ),
+
+                                    
                                   ],
                                 ),
                               ),
